+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Felt::InputGroup::EmailField` component.
 - `Felt::InputGroup::TextField` component.
 - This CHANGELOG file to hopefully serve as an evolving example of a
   standardized open source project CHANGELOG.